# Compute edge adjacency lists lazily (32× speedup, 10× smaller)