--- a/tests/core/test_graph.py
+++ b/tests/core/test_graph.py
@@ -630,13 +630,19 @@ def test_benchmark_get_features(benchmark, num_types, type_arg, feature_size):
 @pytest.mark.parametrize("num_nodes,num_edges", [(0, 0), (100, 200), (1000, 5000)])
 # features or not, to capture their cost
 @pytest.mark.parametrize("feature_size", [None, 100])
-def test_benchmark_creation(benchmark, feature_size, num_nodes, num_edges):
+@pytest.mark.parametrize("force_adj_lists", [False, True])
+def test_benchmark_creation(
+    benchmark, feature_size, num_nodes, num_edges, force_adj_lists
+):
     nodes, edges = example_benchmark_graph(
         feature_size, num_nodes, num_edges, features_in_nodes=True
     )
 
     def f():
-        return StellarGraph(nodes=nodes, edges=edges)
+        sg = StellarGraph(nodes=nodes, edges=edges)
+        if force_adj_lists:
+            sg._edges._init_adj_lists()
+        return sg
 
     benchmark(f)
 
@@ -648,15 +654,19 @@ def test_benchmark_creation(benchmark, feature_size, num_nodes, num_edges):
 @pytest.mark.parametrize("num_nodes,num_edges", [(0, 0), (100, 200), (1000, 5000)])
 # features or not, to capture their cost
 @pytest.mark.parametrize("feature_size", [None, 100])
+@pytest.mark.parametrize("force_adj_lists", [False, True])
 def test_allocation_benchmark_creation(
-    allocation_benchmark, feature_size, num_nodes, num_edges
+    allocation_benchmark, feature_size, num_nodes, num_edges, force_adj_lists
 ):
     nodes, edges = example_benchmark_graph(
         feature_size, num_nodes, num_edges, features_in_nodes=True
     )
 
     def f():
-        return StellarGraph(nodes=nodes, edges=edges)
+        sg = StellarGraph(nodes=nodes, edges=edges)
+        if force_adj_lists:
+            sg._edges._init_adj_lists()
+        return sg
 
     allocation_benchmark(f)