@kieranricardo 2020-04-22
following our slack discussion, I supsect that your gist above might not be measuring the memory of the edge lists and skewing the results!

I made a small edit to explicitly copy over the edge lists before measuring the memory: https://gist.github.com/kieranricardo/536776d95bfbc56e8ab709f1d261cb50

From this gist on FB15K the adj lists are still 60% of the total memory! So I think this PR is still worth while 👍