@huonw 2020-04-27
I think there's an argument to be made either way:

1. the current behaviour measures the incremental memory used by `StellarGraph` over storing the raw dataset in memory, and so answers the question "how much _more_ memory do I need to use StellarGraph over traditional methods"
2. the new behaviour measures something closer to the total memory used for storing the `StellarGraph` long term

However, arguably 2 is better for general user-facing communication, as you say.

Rather than change things here, I'll just wait to #1220, because that changes the edge list format to not be able to reuse the arrays from the external DataFrame.