@kieranricardo 2020-04-28
> Rather than change things here, I'll just wait to #1220, because that changes the edge list format to not be able to reuse the arrays from the external DataFrame.

I'm not 100% on this point, but `pd.concat(type_dfs)` in `ElementData.__init__` duplicates memory of the dataframes so I think we're currently not reusing the arrays from the external dataframe. 

Running this locally and watching my `python` memory usage on activity monitor is ~7Gb.

```
> dfs = [pd.DataFrame(np.ones((1000, 1000))) for _ in range(1000)]
```

Then when I run this my memory usage goes up to 15Gb.
```
> concated = pd.concat(df)
```