@huonw 2020-05-01
I've updated this, and indeed as you suspected in https://github.com/stellargraph/stellargraph/pull/1291#issuecomment-620399157, the iloc work didn't change the behaviour with ownership. I've updated the PR description with the new numbers incorporating the iloc work, and using the version that includes the edge lists https://gist.github.com/kieranricardo/536776d95bfbc56e8ab709f1d261cb50.